reduce complexity and improve redabilty
